backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
beautifulsoup4==4.12.2
cohere==4.5.0
qdrant-client==1.9.0
python-dotenv==1.0.0
openai==1.3.7